--- crates/ghss_models/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "ghss_models"
-version = "0.1.0"
-authors = ["Jan Kuehle <jkuehle90@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-chrono = "0.4.11"
-ghss_github = { path = "../ghss_github" }
-ghss_influxdb = { path = "../ghss_influxdb" }
