+++ crates/ghss_store_client/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+ghss_github = { path = "../ghss_github" }
 prost = "0.6.1"
 tonic = "0.2.1"
