+++ crates/ghss_website/.env.example
@@ -4,10 +4,7 @@ GH_CLIENT_ID=
 GH_CLIENT_SECRET=
 GH_WEBHOOK_SECRET=
 
-INFLUXDB_BASE_URL=
-INFLUXDB_ADMIN_USERNAME=
-INFLUXDB_ADMIN_PASSWORD=
-INFLUXDB_READ_PASSWORD=
+STORE_URL=
 
 TOKEN_SECRET=
