+++ crates/ghss_website/src/config.rs
@@ -10,10 +10,7 @@ pub struct Config {
     pub gh_client_id: String,
     pub gh_client_secret: SecUtf8,
     pub gh_webhook_secret: SecStr,
-    pub influxdb_base_url: String,
-    pub influxdb_admin_username: String,
-    pub influxdb_admin_password: SecUtf8,
-    pub influxdb_read_password: SecUtf8,
+    pub store_url: String,
     pub token_secret: SecStr,
     pub honeycomb_api_key: SecUtf8,
     pub honeycomb_dataset: String,
@@ -33,10 +30,7 @@ pub fn load() -> Config {
         gh_client_id: env("GH_CLIENT_ID"),
         gh_client_secret: SecUtf8::from(env("GH_CLIENT_SECRET")),
         gh_webhook_secret: SecStr::from(env("GH_WEBHOOK_SECRET")),
-        influxdb_base_url: env("INFLUXDB_BASE_URL"),
-        influxdb_admin_username: env("INFLUXDB_ADMIN_USERNAME"),
-        influxdb_admin_password: SecUtf8::from(env("INFLUXDB_ADMIN_PASSWORD")),
-        influxdb_read_password: SecUtf8::from(env("INFLUXDB_READ_PASSWORD")),
+        store_url: env("STORE_URL"),
         token_secret: SecStr::from(env("TOKEN_SECRET")),
         honeycomb_api_key: SecUtf8::from(env("HONEYCOMB_API_KEY")),
         honeycomb_dataset: env("HONEYCOMB_DATASET"),
